# Conflicts:
#   src/main/java/com/nsa/clinical/controllers/OptionController.java
#   src/main/java/com/nsa/clinical/controllers/QuestionController.java
#   src/main/java/com/nsa/clinical/forms/NewOptionForm.java
#   src/main/java/com/nsa/clinical/forms/NewQuestionForm.java
#   src/main/java/com/nsa/clinical/repositories/OptionRepository.java
#   src/main/java/com/nsa/clinical/services/OptionImplementation.java
#   src/main/java/com/nsa/clinical/services/OptionService.java
#   src/main/java/com/nsa/clinical/services/QuestionImplementation.java
#   src/main/java/com/nsa/clinical/services/QuestionService.java
#   src/main/java/com/nsa/clinical/services/QuestionnaireImplementation.java

+++ src/main/java/com/nsa/clinical/entities/Option.java
@@ -17,7 +17,7 @@ public class Option {
     @Id
     @Column(name = "option_id")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer optionId;
+    private Long optionId;
 
     @Column(name = "option_description")
     private String optionDescription;
